Update treechopper.cfg
Removed TG Chainsaw from TreeChopper whitelist as it was assigning a Durability NBT Tag to chainsaws when used. Durability of the chainsaw hitting zero (a stat ignored by TechGuns Chainsaw itself) results in it no longer Tree Chopping. Until we can get a bug fix for this, I'm removing it from the TreeChopper whitelists.